+++ Src/FileTransfer/RequestCurl.h
@@ -11,7 +11,7 @@ class CTaskFileTransfer;
  * \brief libcurl request
  * \note It is running in worker thread
  *
- * Since 2.4.0
+ * \since 2.4.0
  */
 class CRequestCurl : public QObject
 {